+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Breaking changes
 
+* Removed `--no-inline` CLI option and `inline` config option, inline options
+  are now always enabled.
 * Installer script (install.lua) is removed. Luacheck can still be installed
   manually by recursively copying `src/*` to a directory in `package.path`
   and copying `bin/luacheck.lua` to a directory in `PATH` as `luacheck`.